+++ test/fixtures/withServerContext-app-no-getInitialProps/next.config.mjs
@@ -1,9 +1,11 @@
 // @ts-check
 
-/** @type {import("next").NextConfig} */
-export default {
+/** @satisfies {import("next").NextConfig} */
+const nextConfig = {
   pageExtensions: ["mjs"],
   eslint: {
     ignoreDuringBuilds: true,
   },
 };
+
+export default nextConfig;